// src/views/devPatterns/model/Question.enum.ts
export enum QuestionCategoryEnum {
  COMMUNICATION = 'COMMUNICATION',
  CUSTOMER_SERVICE = 'CUSTOMER_SERVICE',
  FOOD_SAFETY = 'FOOD_SAFETY',
  OCCUPATIONAL_SAFETY = 'OCCUPATIONAL_SAFETY',
}

export enum QuestionSortKeyEnum {
  CATEGORY = 'category',
  CODE = 'code',
  CONTENT = 'content',
  STATUS = 'status',
  UPDATED_AT = 'updatedAt',
}

export enum QuestionStatusEnum {
  ACTIVE = 'ACTIVE',
  ARCHIVED = 'ARCHIVED',
  DRAFT = 'DRAFT',
}